+++ BE/todo-12/src/main/java/dev/codesquad/java/todo12/ApiCategoryController.java
@@ -47,22 +47,4 @@ public class ApiCategoryController {
         categoryService.deleteCategory(id);
         return new ResponseEntity(OK, HttpStatus.OK);
     }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    private String catchDataNotFoundException(DataNotFoundException e) {
-        return e.getMessage();
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    private String catchUnauthorizedException(UnauthorizedException e) {
-        return e.getMessage();
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    private String catchInvalidTokenException(InvalidTokenException e) {
-        return e.getMessage();
-    }
 }
