+++ BE/todo-12/src/main/java/dev/codesquad/java/todo12/ApiHomeController.java
@@ -5,9 +5,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -28,23 +26,5 @@ public class ApiHomeController {
     @GetMapping("/history")
     public ResponseEntity history() {
         return new ResponseEntity(historyRepository.findAll(), HttpStatus.OK);
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    private String catchDataNotFoundException(DataNotFoundException e) {
-        return e.getMessage();
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    private String catchUnauthorizedException(UnauthorizedException e) {
-        return e.getMessage();
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    private String catchInvalidTokenException(InvalidTokenException e) {
-        return e.getMessage();
     }
 }
