+++ BE/todo-12/src/main/java/dev/codesquad/java/todo12/ExceptionAdvice.java
@@ -1,0 +1,26 @@
+package dev.codesquad.java.todo12;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+@RestControllerAdvice
+public class ExceptionAdvice {
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    private String catchDataNotFoundException(DataNotFoundException e) {
+        return e.getMessage();
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    private String catchUnauthorizedException(UnauthorizedException e) {
+        return e.getMessage();
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    private String catchInvalidTokenException(InvalidTokenException e) {
+        return e.getMessage();
+    }
+}
