+++ BE/todo-12/src/main/resources/application.properties
@@ -5,17 +5,17 @@ spring.mvc.hiddenmethod.filter.enabled=true
 logging.level.sql=debug;
 logging.level.com.codesquad.qna=debug
 
-# ec2 mySQL 연결
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/todo12
-spring.datasource.username=henry
-spring.datasource.password=henry
-spring.datasource.initialization-mode=always
-spring.datasource.data=classpath:schema.sql, classpath:data.sql
-#spring.datasource.initialization-mode=embedded
+## ec2 mySQL 연결
+#spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+#spring.datasource.url=jdbc:mysql://localhost:3306/todo12
+#spring.datasource.username=henry
+#spring.datasource.password=henry
+#spring.datasource.initialization-mode=always
+#spring.datasource.data=classpath:schema.sql, classpath:data.sql
+##spring.datasource.initialization-mode=embedded
 
-## DB Connection 설정 (로컬 DB 테스트용)
-#spring.datasource.url=jdbc:h2:mem://localhost/~/todo-12;DB_CLOSE_ON_EXIT=FALSE
-#spring.datasource.driverClassName=org.h2.Driver
-#spring.datasource.username=sa
-#spring.datasource.password=
+# DB Connection 설정 (로컬 DB 테스트용)
+spring.datasource.url=jdbc:h2:mem://localhost/~/todo-12;DB_CLOSE_ON_EXIT=FALSE
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
